backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
httpx==0.27.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
jsonschema==4.23.0
tiktoken==0.7.0
python-multipart==0.0.9
sse-starlette==2.1.0
